requirements.txt
websockets
dnspython
selenium
requests
whois
python-Levenshtein